refactor(web-serial-rxjs): unify SerialSessionOptions normalization
Clarify connection vs feature option responsibilities, move buffer
resolve helpers next to their option types, validate bufferSize like
baudRate, and reject invalid limits from createTerminalBuffer.
